polymarket-btc-5m-arb-bot/requirements.txt
requests>=2.31.0
websockets>=12.0
pyyaml>=6.0
python-dotenv>=1.0.0
py-clob-client>=0.18.0
eth-account>=0.10.0
